__syncthreads() after writing shared memory
Threadblocks should sync after writing to shared memory to avoid reading uninitialized values. 
Was rereading this as a reference while trying to help someone with an opengl compute shader question and realized I forgot to sync threads here!
If you've ever noticed inconsistent results between identical runs, this could be a cause. Possible you will notice no difference with this change, but definitely a good idea to include this in future runs.