Reduce internal copies
Changed a bunch of pass by values to pass by (const) reference so that the program is not creating a new copy of the data structures with each function call.

This also marks a bunch of functions as const (where applicable) to allow the const references.

Simple benchmark tests show 10% - >50% speedup for some of the functions.

Added tests to cover the changes.