GA resets environment variables making CUDA and .local/bin undetectable.
The git version availible on Baldur at the moment is too old to be
used with the --recursive option. This breaks the GA checkout
action, which defaults to downloading the repo using the GitHub's
REST API.
The cloned repo does not contain the .git directory, i.e. the
cloned files are not an actual git repository. This breaks any
action that relies on git commands to deploy/checkout/etc,
including initializing and updating submodules. Without
cloning pybind11 we can't build.
Additionally, still unclear how to correctly set the
environment to detect correct CUDA compiler. This breaks
build.