Works on AWS, issue is in env setup on Baldur.
Removed Baldur from list of GA Runners, tested it on AWS. Worked
Difference is that system Python and Git can be altered on AWS,
but not on Baldur. Means that the ways I set up the environment
are incorrect.
The setup-conda action removes all .profiles, .*rc files and more.
This makes it impossible to shortcut things by altering PATH.