Splitting bindings into multiple files.
Proposal on how to split Pybind bindings into multiple files for
readability. Implemented on the example of a PSF class.

The code currently doesn't work because we can not have multiple
entry point calls to the macro PYBIND11_MODULE in a single codebase.
Or at least I don't think we can and nothing I tried worked.
We certainly can not extend a once-declared module by repeatedly
calling PYBIND11_MODULE, but seemingly we can not define a new one
either.