Normalize StackSearch class method naming.
KBMOSearch has been renamed StackSearch to match Python bindings.
I have no preference regarding the naming of the class, it can
rename KBMOSearch - as long as it's consistent with its Python
bindings.

In general no other class had method names as confusing or as
disparate from its Python bindings so in this merge, I've
renamed multiple of its methods - which was not the case for
other classes in this refactor. The following renames
occured:

* get_image_stack -> get_imagestack
* scienceStampsForViz, science_viz_stamps -> get_stamps
* [science, sci, mean, median, coadd, summed]_[sci, science]_stamp[s] -> [mean, median, coadd, summed]_stamp[s]
* get_traj_pos, getTrajPos -> get_trajectory_position
* get_mult_traj_pos, getMultTrajPos -> get_trajectory_positions
* [psi, phi]_curves -> get_[psi, phi]_curves

Same like with the class name, I'm not married for either particualar
individual variation on the topic, but the **only** kind of indirection
that should occur in the bindings is to make the interface feel and
act more Pythonic, by for example wrapping [get, set]_method using a
pybind11::def_property and then not exposing the individual getters
and setters, or for example, wrapping print, stringify and similar
methods in the dunder str, dunder repr methods and then also not
exposing the internals in the binding code.